+++ pkg/skaffold/color/formatter.go
@@ -28,6 +28,40 @@ import (
 // for testing to an arbitrary method.
 var IsTerminal = isTerminal
 
+// Color can be used to format text using ANSI escape codes so it can be printed to
+// the terminal in color.
+type Color int
+
+var (
+	// LightRed can format text to be displayed to the terminal in light red, using ANSI escape codes.
+	LightRed = Color(91)
+	// LightGreen can format text to be displayed to the terminal in light green, using ANSI escape codes.
+	LightGreen = Color(92)
+	// LightYellow can format text to be displayed to the terminal in light yellow, using ANSI escape codes.
+	LightYellow = Color(93)
+	// LightBlue can format text to be displayed to the terminal in light blue, using ANSI escape codes.
+	LightBlue = Color(94)
+	// LightPurple can format text to be displayed to the terminal in light purple, using ANSI escape codes.
+	LightPurple = Color(95)
+	// Red can format text to be displayed to the terminal in red, using ANSI escape codes.
+	Red = Color(31)
+	// Green can format text to be displayed to the terminal in green, using ANSI escape codes.
+	Green = Color(32)
+	// Yellow can format text to be displayed to the terminal in yellow, using ANSI escape codes.
+	Yellow = Color(33)
+	// Blue can format text to be displayed to the terminal in blue, using ANSI escape codes.
+	Blue = Color(34)
+	// Purple can format text to be displayed to the terminal in purple, using ANSI escape codes.
+	Purple = Color(35)
+	// Cyan can format text to be displayed to the terminal in cyan, using ANSI escape codes.
+	Cyan = Color(36)
+	// None uses ANSI escape codes to reset all formatting.
+	None = Color(0)
+
+	// Default default output color for output from Skaffold to the user
+	Default = Blue
+)
+
 func wrapTextIfTerminal(out io.Writer, c Color, a ...interface{}) string {
 	if IsTerminal(out) {
 		return c.Sprint(a...)
@@ -35,27 +69,41 @@ func wrapTextIfTerminal(out io.Writer, c Color, a ...interface{}) string {
 	return fmt.Sprint(a...)
 }
 
-// Fprint wraps the operands in the color ANSI escape codes, and outputs the result to
+// Sprint will format the operands such that they are surrounded by the ANSI escape sequence
+// required to display the text to the terminal in color.
+func (c Color) Sprint(a ...interface{}) string {
+	text := fmt.Sprint(a...)
+	return fmt.Sprintf("\033[%dm%s\033[0m", c, text)
+}
+
+// Sprintf will format the operands according ot the format specifier and wrap the resulting text
+// with the ANSI escape sequence required to display the text to the terminal in color.
+func (c Color) Sprintf(format string, a ...interface{}) string {
+	formatSpecifier := c.Sprint(format)
+	return fmt.Sprintf(formatSpecifier, a...)
+}
+
+// Fprint wraps the operands in c's ANSI escape codes, and outputs the result to
 // out. If out is not a terminal, the escape codes will not be added.
 // It returns the number of bytes written and any errors encountered.
-func Fprint(out io.Writer, c Color, a ...interface{}) (n int, err error) {
+func (c Color) Fprint(out io.Writer, a ...interface{}) (n int, err error) {
 	t := wrapTextIfTerminal(out, c, a...)
 	return fmt.Fprint(out, t)
 }
 
-// Fprintln wraps the operands in the color ANSI escape codes, and outputs the result to
+// Fprintln wraps the operands in c's ANSI escape codes, and outputs the result to
 // out, followed by a newline. If out is not a terminal, the escape codes will not be added.
 // It returns the number of bytes written and any errors encountered.
-func Fprintln(out io.Writer, c Color, a ...interface{}) (n int, err error) {
+func (c Color) Fprintln(out io.Writer, a ...interface{}) (n int, err error) {
 	t := wrapTextIfTerminal(out, c, a...)
 	return fmt.Fprintln(out, t)
 }
 
 // Fprintf applies formats according to the format specifier (and the optional interfaces provided),
-// wraps the result in the color ANSI escape codes, and outputs the result to
+// wraps the result in c's ANSI escape codes, and outputs the result to
 // out, followed by a newline. If out is not a terminal, the escape codes will not be added.
 // It returns the number of bytes written and any errors encountered.
-func Fprintf(out io.Writer, c Color, format string, a ...interface{}) (n int, err error) {
+func (c Color) Fprintf(out io.Writer, format string, a ...interface{}) (n int, err error) {
 	var t string
 	if IsTerminal(out) {
 		t = c.Sprintf(format, a...)
