+++ pkg/skaffold/color/formatter_test.go
@@ -35,13 +35,29 @@ func compareText(t *testing.T, expected, actual string, expectedN int, actualN i
 	}
 }
 
+func TestColorSprint(t *testing.T) {
+	c := Red.Sprint("TEXT")
+	expected := "\033[31mTEXT\033[0m"
+	if c != expected {
+		t.Errorf("Expected %s. Got %s", expected, c)
+	}
+}
+
+func TestColorSprintf(t *testing.T) {
+	c := Green.Sprintf("A GREAT NUMBER IS %d", 5)
+	expected := "\033[32mA GREAT NUMBER IS 5\033[0m"
+	if c != expected {
+		t.Errorf("Expected %s. Got %s", expected, c)
+	}
+}
+
 func TestFprint(t *testing.T) {
 	orgIsTerminal := IsTerminal
 	defer func() { IsTerminal = orgIsTerminal }()
 	IsTerminal = func(_ io.Writer) bool { return true }
 
 	var b bytes.Buffer
-	n, err := Fprint(&b, Green, "It's not easy being")
+	n, err := Green.Fprint(&b, "It's not easy being")
 	expected := "\033[32mIt's not easy being\033[0m"
 	compareText(t, expected, b.String(), 28, n, err)
 }
@@ -52,7 +68,7 @@ func TestFprintln(t *testing.T) {
 	IsTerminal = func(_ io.Writer) bool { return true }
 
 	var b bytes.Buffer
-	n, err := Fprintln(&b, Green, "2 less chars!")
+	n, err := Green.Fprintln(&b, "2 less chars!")
 	expected := "\033[32m2 less chars!\033[0m\n"
 	compareText(t, expected, b.String(), 23, n, err)
 }
@@ -63,7 +79,7 @@ func TestFprintf(t *testing.T) {
 	IsTerminal = func(_ io.Writer) bool { return true }
 
 	var b bytes.Buffer
-	n, err := Fprintf(&b, Green, "It's been %d %s", 1, "week")
+	n, err := Green.Fprintf(&b, "It's been %d %s", 1, "week")
 	expected := "\033[32mIt's been 1 week\033[0m"
 	compareText(t, expected, b.String(), 25, n, err)
 }
@@ -75,7 +91,7 @@ func TestFprintNoTTY(t *testing.T) {
 
 	var b bytes.Buffer
 	expected := "It's not easy being"
-	n, err := Fprint(&b, Green, expected)
+	n, err := Green.Fprint(&b, expected)
 	compareText(t, expected, b.String(), 19, n, err)
 }
 
@@ -85,7 +101,7 @@ func TestFprintlnNoTTY(t *testing.T) {
 	IsTerminal = func(_ io.Writer) bool { return false }
 
 	var b bytes.Buffer
-	n, err := Fprintln(&b, Green, "2 less chars!")
+	n, err := Green.Fprintln(&b, "2 less chars!")
 	expected := "2 less chars!\n"
 	compareText(t, expected, b.String(), 14, n, err)
 }
@@ -95,7 +111,7 @@ func TestFprintfNoTTY(t *testing.T) {
 	IsTerminal = func(_ io.Writer) bool { return false }
 
 	var b bytes.Buffer
-	n, err := Fprintf(&b, Green, "It's been %d %s", 1, "week")
+	n, err := Green.Fprintf(&b, "It's been %d %s", 1, "week")
 	expected := "It's been 1 week"
 	compareText(t, expected, b.String(), 16, n, err)
 }
